Add new Web Reputation API client
