--- src/index.ts
@@ -1,6 +1,0 @@
-import FastCache from './fastcache';
-import MCSkinHistory from './mcskinhistory';
-import VATAPI from './vat';
-import WebReputationAPI from './web-reputation';
-
-export { FastCache, MCSkinHistory, VATAPI, WebReputationAPI };
